fix multiple bugs: thread safety, null pointers, command injection
- Fix use-after-free in preview_cache.cpp: replace detached threads with
  managed threads that are properly joined on destruction
- Fix missing null-termination in file_browser.cpp strncpy call
- Fix unhandled exception in preview_cache.cpp getFileHash() when file
  doesn't exist
- Fix command injection vulnerability in app.cpp openInBlender/openContainingFolder
  by escaping shell special characters
- Fix potential integer overflow in blend_parser.cpp thumbnail size calculation
- Fix null pointer dereference from sqlite3_column_text in database.cpp by
  adding safeColumnText() helper
- Fix lock order deadlock in thumbnail_cache.cpp processLoadedThumbnails by
  restructuring to avoid holding both mutexes simultaneously

+++ src/blend_parser.cpp
@@ -81,7 +81,9 @@ std::optional<BlendThumbnail> BlendParser::extractThumbnail(std::ifstream& file,
         return std::nullopt;
     }
 
-    size_t pixelDataSize = static_cast<size_t>(width * height * 4);
+    // Perform multiplication in size_t to prevent potential overflow
+    // (even though current limits prevent it, this is defensive)
+    size_t pixelDataSize = static_cast<size_t>(width) * static_cast<size_t>(height) * 4;
     if (block.size < static_cast<int32_t>(8 + pixelDataSize)) {
         return std::nullopt;
     }
